+++ CHANGELOG.md
@@ -7,10 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 - Support passing additional args to all change events
 - Allow "setting" list properties via the `property.value()` method
+- Get rid of special case for list property default rules to run on init new
+- Get rid of unused event tracking code in the `updateArray` function
 ### Fixed
 - Dont' ignore null value for reference property in initial entity state
 - Improve type checking of property options (i.e. `identifier` option)
 - Remove duplicate exports of 'initilization-context'
+- Always set pending init to `true` when a non-constant property initializes without an initializer,
+  and ensure that it is consistently set to `false` when the property is set, the array is modified,
+  or the property is included in the entity's initial state
 ### Added
 - Added ES module build output
 ## [0.8.42] - 2023-01-26